requirements.txt
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.38.0
certifi==2017.7.27.1
chardet==3.0.4
Click==7.0
crcmod==1.7
cycler==0.10.0
cymem==1.31.2
Cython==0.26
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
en-core-web-sm==1.2.0
enum34==1.1.6
Flask==1.0.2
ftfy==4.4.3
funcsigs==1.0.2
functools32==3.2.3.post2
google-compute-engine==2.2.4
h5py==2.7.0
html5lib==0.999999999
idna==2.6
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.0.6
lime==0.1.1.23
Markdown==2.6.8
MarkupSafe==1.1.1
matplotlib==2.0.2
mock==2.0.0
murmurhash==0.26.4
networkx==1.11
nltk==3.2.4
numpy==1.16.2
olefile==0.44
opencv-python==3.3.0.10
pathlib==1.0.1
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.3.0
ptyprocess==0.5.2
pycocotools==2.0.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
regex==2017.7.28
requests==2.18.4
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
spacy==1.9.0
subprocess32==3.2.7
tensorflow==1.1.0
termcolor==1.1.0
Theano==0.9.0
thinc==6.5.2
toolz==0.8.2
tqdm==4.15.0
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.10.11
